[BEAM-1871] Move out test only utilities used only in examples/java out of sdks/java/core.
