+++ examples/java/src/test/java/org/apache/beam/examples/WordCountIT.java
@@ -20,10 +20,10 @@ package org.apache.beam.examples;
 
 import java.util.Date;
 import org.apache.beam.examples.WordCount.WordCountOptions;
+import org.apache.beam.examples.testing.FileChecksumMatcher;
 import org.apache.beam.sdk.io.FileSystems;
 import org.apache.beam.sdk.io.fs.ResolveOptions.StandardResolveOptions;
 import org.apache.beam.sdk.options.PipelineOptionsFactory;
-import org.apache.beam.sdk.testing.FileChecksumMatcher;
 import org.apache.beam.sdk.testing.TestPipeline;
 import org.apache.beam.sdk.testing.TestPipelineOptions;
 import org.junit.BeforeClass;
